feat(prometheus): New PromQL Parser (#184)
* First cut PromQLParser

* Adding test cases

* Merging latest

* Query to LogicalPlan basic

* Working tests and changes in LP

* Moving to a separate module

+++ query/src/main/scala/filodb/query/LogicalPlan.scala
@@ -46,7 +46,7 @@ case class RawSeries(rangeSelector: RangeSelector,
   */
 case class PeriodicSeries(rawSeries: RawSeries,
                           start: Long,
-                          step: Int,
+                          step: Long,
                           end: Long) extends PeriodicSeriesPlan
 
 /**
@@ -58,7 +58,7 @@ case class PeriodicSeries(rawSeries: RawSeries,
   */
 case class PeriodicSeriesWithWindowing(rawSeries: RawSeries,
                                        start: Long,
-                                       step: Int,
+                                       step: Long,
                                        end: Long,
                                        window: Int,
                                        function: RangeFunctionId,
